wcmUsb: use struct input_absinfo instead of an integer array.
kernel 2.6.31 introduced a resolution field in the absinfo struct, thus the
nValues[5] array wasn't large enough. This again leads to a memory overrun
and random crashes on init.

Instead of increasing nValues, replace it with input_absinfo which is the
correct struct to pass in anyway. Do the same for the sID array (of type
struct input_id).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>