Don't overwrite the name with the XI type.
Since the local device is now the actual device, not the fake local, we
don't need to set the name here. Also, the name is the identifier, I don't
know if it ever actually was used for the XI defined names (wich are the
type_name instead).

This has the potential to break clients that were looking for the "Wacom
erasor" type name instead of the XI defined types ("ERASOR"). If so, a
simple change fixes this.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>