Support auto-hotplugging in the driver.
If a device is added without the "Type" option set (a previously invalid
configuration), it is assumed to be hotplugged. Check the list of known
products for the supported tool types and init a new device for each tool
type.

All devices added as a group have a 'uniq' identifier (current timestamp).
If the parent device is being removed, check for still-existing devices with
the same uniq and remove those too.

The server closes devices in-order, since the dependent devices are added
before the parent, this is a noop during CloseDownDevices.

The .fdi file doesn't to merge the type now.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>