If a device is specified, don't check for a wacom vendor id.
Testing for the vendor ID is useful when using auto-dev probe (though it
might return false negatives). If a user was capable enough to specify a
device option, assume they know what they're doing and just try to init the
device.

This ensures that the wacom driver works against serial devices that don't
return a vendor ID. And it ensures that the driver can be used with other
tablets by different vendors but a similar behaviour.

Reported-by: Thomas Jaeger
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>