Init the key class struct for the device (and send keys) on ABI 7.
Force a hardcoded keymap to semi-protect against local keymap changes (i.e.
other layouts). This doesn't protect against run-time changes of course, so
it's arguably pointless if we expect the session to set the RMLVO.

This patch restores key sending for the tablets as working before the switch
to input ABI 7 removed that bit.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>