Convert a failed ioctl error into a warning only.
With hot-plugged devices, we don't know for sure if the device is an USB or
serial, so we default it to USB. If the device happens to be a serial
device, the ioctl fails with EIO and we try to add it as serial device.
Hence a warning, not an error.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>