Remove ForceDevice option - now unused.
Since we only support ISDV4 serial devices with this driver, there is no
need to check the "ForceDevice" option any more, which was used to
distinguish ISDV4 from regular serial devices.

We check if the device is a serial device or not by accessing its file
descriptor directly.

Device Class is assigned to USB by default.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>