AC_TYPE of 0 is a valid type for wheel events.
3cfc3f5df754c6266127e3addbd6ee84c8f471d9 introduced button actions. Wheel
events by default are simply assigned to a button. In the old code,
AC_BUTTON was 0, now AC_BUTTON is a specific event mask. Hence we need to
cater for both AC_BUTTON set or unset - they just happen to produce the same
result.

This is a quickfix, the wheel event processing seems it could use some
cleanup.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>