Set Touch and Gesture options during the configuration
We moved the touch and gesture option check from usbWcmInit
to wcmValidateDevice.c since usbWcmInit gets called every time
the device is opened, which is unnecessary for option check
in most cases.

The Tablet PC Button option should also be dealt in the same way.
But it requires extra defines to make it clean, which will affect
more files in the driver. We'll do it in a separate patch later.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>