Remove area overlap check for area property
Temporarily change the area to the new coordinates, do the range overlap
check and then change it back to the original. This is needed since
otherwise the area overlap check always fails (since the new area will be
mostly identical to the current one).

New behaviour: all four coordinates set to -1 reset the area to the
defaults.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>