Auto-prepend device name/file to debug messages.
Add a new "name" field to the WacomDeviceRec. This field is at the same
offset as the wcmDevice Field of the WacomCommonRec, hence the access to
either is the same if casted.

Then use this field to auto-prepend either the device name or the device
file wherever a DBG macro is used. The output for DBG on priv is now
(II) Wacom Intuos4 6x9 (xf86WcmMappingFactor): X factor = 0.0258, Y factor = 0.0309

and for DBG on common
(II) /dev/input/event13 (wcmReadPacket): buffer has 112 bytes

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>