Add "Wacom Debug Levels" property.
This property is only present if compiled with --enable-debug and allows for
run-time modification of the debug levels. Values supported are 0 to 10,
with 10 being the most verbose.

xsetwacom --set "device name" DebugLevel 10
xsetwacom --set "device name" CommonDBG 10

Since this is a driver-internal property clients are discouraged from using
this property for any purpose. It may disappear or change without warning in
future releases of the driver.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>