Update README to reflect current state a bit better.
This driver is an X11 driver only, and the configure options are explained
by configure --help. Let's not mirror this in the README, it may diverge
otherwise.

Add a very simple explanation on how to set up the driver fro hotplugging
with HAL, xorg.conf configuration is better explained in the man page.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>