config: move compiler flags to src/Makefile.am
The -fno-strict-aliasing option was added in CWARNFLAG some times ago. Aside
from the issue that it isn't a warning flag, being in this macro had the
side-effect of affected all modules when just a few needed it.

The plan is to transfer this option back to the module that are actually
using CWARNFLAG. Each module can then decide if they really want it or not.
This patch is just prep work.

The net change will be that strict aliasing optimization will be performed
(by default in -O2) and -Wstrict-aliaing=2" will be added to flag suspicious
code. This has not been submitted for review yet.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>