Take scroll buttons into accounts when allocating buttons.
X expects buttons 4-7 to be the scrolling directions, hence if our device
has more than 3 buttons, we need to add 4 to the actual number of buttons
reported. If the device has 3 buttons only, use 7 anyway so we can scroll.

This affects the internal default button map as well, physical buttons need
to be offset by 4 after the first three.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>