Only allow absolute mode for pads.
The driver doesn't really know anymore whether a device is a core device or
not since it may change at any time after an XIChangeHierarchy() call. Even
then, a device is never a core device anyway - the master device it is
attached to is.

This code can't be triggered in server 1.7 or higher anyway, just purge it
and return an error when trying to set the pad to relative mode.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>