Remove user setting of MaxX/MaxY/MaxZ in xorg.conf.
Any values specified by user in xorg.conf for MaxX/Y/Z
and the MaxTouch* versions were always being overwritten
during call to device->GetRanges().

This is a cleanup to remove this broken feature since users
really should not need to touch this.  Removing is much
easier then getting user override values to work correctly
for values that are set so late.

This broken interface was not document to user in wacom man pages
and users are more interested in BottomX/Y and pressure curves
then the hopefully rare case of trying to debug new hardware.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by:  Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>