Open the fd only once during PreInit.
Instead of having every function open and close the fd, just open it once
towards the start of PreInit and then close it at the end again. Anything it
between can just re-use it.

This reshuffles some of the code in PreInit that was supposedly working when
"auto-dev" was set. It wasn't, so this just emphasises the brokenness.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>