Revamp device PreInit process.
The new init process:
- allocate the generic struct used by all device types.
- if no device is given, auto-probe for one (find a wacom device
in /dev/input/event?
- open the device file
- probe the device

The allocation process is different now, the common settings are allocated
for all devices, type-specific settings are set once some information about
this device is available.

The auto-dev probe is now invoked for each device that does not have the
device option set. For this to work, we need the common->wcmDevice to be
initialized to NULL.  An option "Device" "auto-dev" is _not_ supported, it
would be interpreted as file path and fail. Something for the future maybe.

Probing could still be improved by getting more information from the device
now instead of during open. Something else for the future.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>