Fix usbGetRanges() for any load order of touch/pad
For Bamboo P&T devices, both a touch and pad type are
shared for single input device and can be called in any
order based on xorg.conf.  And normal HAL order is to
invoke usbGetRanges() from pad first.

This meant all logic related to IsTouch() was not
correctly running.  Updated to treat pad type as
touch in special case of Bamboo P&T.

This is applies to devices that have both BTN_TOOL_FINGER
and BTN_TOOL_DOUBLETAP which currently is limitted to
Bamboo touch devices.

Signed-off-by: Chris Bagwell <chris@cnpbagwell>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by:  Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>