ISDV4: use wcmSkipInvalidateBytes from wcmSerialValidate.
Untangle the double if inside the loop to perform the same functionality:
first check the header byte and if it has the magic bit set. If not, skip
along to the next byte that does.
Then, check that the remaining bytes don't have the magic bit set. If not,
skip up to the first byte that does have the header bit set.

This patch also changes the function signature so we can print the name of
the device where it happens.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>