Always put the pad in relative mode.
Fixes Intuos3 scrolling issues.

The pad cannot be in absolute mode as it sends the axis values to the
server. Since the pad never gets x/y coordinates from the tablet the server
will fill in the defaults (0/0) for it - even if first_valuator is always >
1. This results in the pointer being reset to the screen origin each time
the pad's scroll strip is used.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>