Normalize pressure sensitivity
Instead of reporting the raw pressure, the normalized pressure from
0 to FILTER_PRESSURE_RES (which is 2048) is reported. This is mainly
to deal with the case where heavily used stylus may have a "pre-loaded"
initial pressure. This patch checks the in-prox pressure and subtract
it from the raw pressure to prevent a potential left-click before the
pen touches the tablet.

Left click threshold and pressure curve are updated accordingly.

Signed-off-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>