Adjust N-Trig pattern.
The old pattern is the default name set by hid and along with some other
changes I added descriptive names for the event devices.  For people using
the older versions of the kernel driver, they are probably better off just
using the wacom driver to support both pen and touch.  For the current
versions of the driver, there isn't much of an advantage to continue using
wacom for touch, and some distinct advantages to moving to evdev
(particularly as MT support matures). The wacom driver is still better for
the pen.

Signed-off-by: Rafi Rubin <rafi@ugcs.caltech.edu>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>