Don't add the OutOfProximity flag to the axis mode.
The server and many clients tends to check for Absolute and/or Relative, but
don't use Absolute as a flag. The XI protocol spec only allows for 0 or 1
(Relative or Absolute), not for an extra flag.

As a result, a device set up with (Absolute|OutOfProximity) will be
interpreted as Absolute by some clients and relative by others.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <pinglinux@gmail.com>