Add two finger gestures for Bamboo with touch
Both src and tools/include are updated. This "feature" will be removed
when gesture is supported in X server or in desktop environment.

Peter Hutterer's comment:

"just for the record: I am deeply opposed to implementing gestures in the
driver. This way every driver needs to cook their own soup, there is hardly
any possibility to let clients configure the gestures, clients won't know
when a gesture has been applied or when a gesture should be skipped instead,
there is virtually no chance of standardization of gestures as all drivers
will end up using slighlty different parameters, etc. the list is rather long.
that's also why I kinda pretend that the code doesn't exist and do cursory
checks only (I can't test it anyway).

having said that, I realise that the lack of gesture support in the server
is currently blocking any other attempts. So I'll reluctantly keep saying
yes to this gesture approach, but expect it to be temporary only."

Signed-off-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>