xsetwacom: remove unimplemented NumScreens setting.
AFAICT, NumScreens in linuxwacom returns screenInfo.numScreens - the same
value as reported by ScreenCount(3). If this is 1 and TwinView is enabled,
this value is hardcoded to 2. This logic can be replicated in the client,
there's no need to export this from xsetwacom. The driver does not export
this anymore anyway - was unimplemented since the fork.

Note: the driver does not export this setting as input driver do not hook
into the RandR notification mechanisms. The input driver thus would not be
notified when a new screen is plugged in, resulting in stale values being
stored in the driver.

Note: this logic is incorrect if RandR 1.2 drives multiple screens.
Although there are two physical screens, there is only one protocol screen.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>
Acked-by: Ping Cheng <pinglinux@gmail.com>