Don't pass the axes around to sendCommon() and friends.
The only events generated here are button events, there's no need to send
the axis values around.

This also fixes a bug where axis values are sent twice to the X server:
The following call order is possible in wcmSendEvents
xf86PostMotionEvent(..., axisvalues[])
-> sendCommonEvents()
    xf86PostButtonEvent(..., axisvalues[])

This is fine for absolute mode but in relative mode this would result in the
axis values applied twice. By not passing the values down, this bug is
avoided.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <pinglinux@gmail.com>