config: wacom depends on RANDR through xf86.h
The xf86.h file contains external declarations specific to RANDR
If this extension is defined (installed) wacom will not compile
if randrproto is not installed under the same prefix.
Because this is an optional extension, XORG_DRIVER_CHECK_EXT
must be used.

move inputproto check to PKG_CHECK_MODULES:
The XORG_DRIVER_CHECK_EXT macro conditionally adds inputproto
to PKG_CHECK_MODULES if this extension is defined in xorg-server.h.
It allows a module to use ifdef XINPUT if the extension is available.
In the case of inputproto, it is part of server and always defined.
In this module, the extension is required and not optional.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>