Don't reconfigure the axes when switching mode.
Run-time axis configuration is not really allowed. The XI2 protocol permits
it but it's forbidden for XI1 clients since they have no notification
mechanisms. And we don't have any server APIs for the XI2 stuff yet.

If a client issues ListInputDevices and the driver reconfigures _after_
that, the client assumes wrong coordinate ranges, causing possible scaling
issues. The right way to handle axis reconfiguration is to scale _inside_
the driver only.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <pinglinux@gmail.com>