Remove wcmSetScreen(), it's not working.
The current behaviour on a single _protocol_ screen is working fine (this
includes RandR 1.2 multi-monitor setups).

The behaviour on a multi-ple protocol screen is broken, with the center area
of the screens being unaccessible. Given two monitors in a left/right
layout, the pointer will jump half-way through the first screen to the
center of the right screen and continue from there. The areas marked with x
in the diagram below are inaccessible.

+-----------+ +-----------+
|     xxxxxx| |xxxxx      |
|     xxxxxx| |xxxxx      |
+-----------+ +-----------+

Remove the wcmSetScreen() call and let the X server handle the screen
crossing. This isn't quite perfect yet as the pointer gets stuck on the
rightmost screen but it's better than the current behaviour.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>