Revert "Don't add the OutOfProximity flag to the axis mode."
This reverts commit 060ad78cf464508f1d123c00b7340838a27f19e3.
Error on my side, this mistake was caused by a bug in the X server code when
listing XI2 devices.

According to the XI1 protocol spec, the device mode for a ListInputDevices
request may only be Relative or Absolute, the mode obtained from a
QueryDeviceState request may be binary OR'd with OutOfProximity. The server
uses the same field (dev->valuator->mode) to store this information but
returns only the LSB from ListInputDevices. The XI2 code missed this part,
returning the mode field as a whole, thus sometimes causing clients to
misinterpret the mode field.

Conflicts:

	src/xf86Wacom.c

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <pinglinux@gmail.com>