Revert "Don't pass the axes around to sendCommon() and friends."
This patch (the one reverted hereby) had an interesting side-effect.
Since the valuators were passed down to the button event, they were included
in the XI events. Thus any client listening to button events would also
obtain the current coordinates with that event. Once the valuators weren't
passed down anymore, there was no way for clients to get the current state
of the valuators with the event itself, they'd have to listen to motion
events and interpolate from that the positions during the button event.

Patch reverted, so that button events include the coordinates again.

This reverts commit 579f31c628f0b0036f3c238cc7f32ccfc32adab1.

Conflicts:

	src/wcmCommon.c

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <pinglinux@gmail.com>