Don't mangle proximity state for the pad.
When the pad is relative (i.e. always, since we force it to), the first
value from the abswheel is stored in priv->oldWheel and then subtracted from
the current state - leading to a v5 of 0. That again leads to a dead wheel,
since we never get any values but 0 from the wheel.
The pad provides proximity values anyway, so don't hack around it
- accept what the hardware says.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Ping Cheng <pinglinux@gmail.com>