Split model probing, DEVICE_INIT and DEVICE_OPEN into separate functions.
Probe for the model during PreInit and get the important bits off the device
here.
During DEVICE_INIT, just register the X server bits required for the device
to be a device.
DEVICE_OPEN actually opens the device, but by then we already know
everything about it anyway.

A large hunk of this patch is just whitespace change due to a indentation
level change in wcmRegisterX11Devices.

This also removes the need for wcmDevOpenCount and wcmInitKeyClassCount.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>