Purge ForceDevice option - unneeded.
We're smart enough to detect if the device is running off a serial port or
off a USB port. The only thing we did with the ForceDevice option was to set
it once (based on a serial ioctl) and then check for it later during the
BaudRate xorg.conf option parsing.

Get rid of the field, parse the BaudRate for all devices even though we
don't actually care about it in the USB code. So setting an invalid baud
rate for a USB device will fail the device init but then again - don't do
that.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>