Only execute GetRanges() once for ISDV4 devices.
ISDV4 tablets don't seem to like multiple query commands in a row, the reply
for those are garbage. Hence, put a barrier in to prevent this. Since the
isdv4data is shared between all devices off the same port, this simply skips
the work of the isdv4GetRanges() and returns early with success.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>