ISDV4: don't start the tablet until the last device is enabled.
Only send one SAMPLING command to the tablet for the last device. If
sampling is started before the last device was enabled, that device may not
work.

This is a quickfix, there is bound to be a better way to solve this but
meanwhile we might as well try to make the tablet work again.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Ping Cheng <pinglinux@gmail.com>