Open the device as serial device in PreInit.
A simple open(2) call works fine for USB devices but for serial devices
the right port parameters need to be set. By luck, sometimes these may
be correct and the device works, other times it won't, leading to
unpredictable serial tablet behaviour.

wcmOpen() already does this for us correctly with xf86OpenSerial().
Re-use it to get the right port parameters. xf86OpenSerial() ignores
serial parameters for non-serial devices, thus it is safe to use on USB
devices as well.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Ping Cheng <pinglinux@gmail.com>