Don't load keysyms into the driver, use keycodes instead.
This addresses a hang in the driver when buttons are configured to send
keystrokes. The current code calls XkbGetCoreMap() which allocates during
the signal handler, causing server hangs.

This patch changes the driver to use keycodes instead of keysyms. There are
a number of reasons for this:
- XkbCoreMap() only hands us a core keymap, not the XKB one that everyone
  uses these days. There are some differences between the core and the XKB
  map (mainly the use of modifiers), hence we should be focusing on the XKB
  map anyway.
- Parsing the XKB map manually in the driver is complicated and not
  something we want to do, especially when trying to maintain ABI compat
  with multiple X server versions. On the other hand, the client-side API
  for XKB is frozen like all of Xlib.
- The driver is not notified of XKB layout changes, thus a cached map would
  potentially be out of date. Moving keysym to keycode conversion to the
  client avoids this.
- The server does not communicate through keysyms to any client, keycodes
  are the only information. It is up to the client to pick the keysym (and
  glyph) based on the keycode.

Note that this changes the property API. Clients that would write straight
to a button mapping property will need to change. AFAICT, only xsetwacom
does this at this point and xsetwacom is updated with this patch.

Note that the current xsetwacom implementation will only focus on level 0
keys. It will not automatically fill in the modifier states. Thus, to get
e.g. a string of "aBc", the following configuration is required:
    xsetwacom set "device" "Button1" "key a +shift B -shift c"

xsetwacom releases modifiers at the end of the configuration, thus to get a
single uppercase letter (or multiple letters), the following is enough:
    xsetwacom set "device" "Button1" "key shift A"

Note: this breaks Ctrl+/- zoom support for keyboard layouts that have +/- on
a different key than the 'us' keyboard layout. Did I mention that hacking up
gestures in the driver is a bad idea?

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>