Make 19200 the default baud rate for 0x90 TPCs
This implements using 19200 as a default baudrate for some tablets
instead of always using 38400.

The baudrate switching logic in isdv4GetRanges is extended to be able
to switch to 38400 as well. It also now outputs a warning.

My WACf004 needs a baudrate of 19200, and is included in the 0x90
tablet_id. This seemed like the easiest way, but I'm unsure whether
all 0x90 tablets use 19200.

Signed-off-by: Jan Steffens <jan.steffens@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>