xsetwacom: add "MapToOutput" parameters.
Provides a new xsetwacom parameter to restrict the tablet to a given RandR
1.2 output. The output has to be active. The command

  xsetwacom set <device name> "MapToOutput" VGA1

maps the coordinates of the tablet to the given output. This is a once-off
effect only. xsetwacom does not track the output for reconfiguration. If the
output is reconfigured, re-running the command will re-configure the tablet
with the new configuration. A list of outputs can be obtained with e.g.
xrandr.

If used together with rotation, the tablet must be rotated before being
mapped to the output.

Requires the "Coordinate Transformation Matrix" property to be present and
handled by the X server.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>