Purge TwinView driver support.
TwinView is a video driver feature that, quite frankly, screws everyone. It
exposes only a single ScreenRec but we need to pretend we have 2 when it's
activated. We can't get the dimensions of the screen and rely on the user to
set it, hoping that everything works in spite of that.

For single ScreenRec setups (including RandR multiscreen), we can use the
Coordinate Transformation Matrix property to achieve tablet-to-screen
mappings (see commit 239822f9f4 "xsetwacom: add "MapToOutput" parameters.")

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>