Remove custom keymap assignment.
Any client that applies keymaps after the device was detected will overwrite
this custom keymap. This is the case with e.g. GNOME, so going through the
effort of defining a keymap is short-lived.

If a special keymap is required, the XKB options can be set accordingly
though they will suffer from the same issue described above.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>