Allow valid case of 2 events per sync windows
The check to find event sync windows with to little information
was to agressive.  For example, if only pressure is changing
then only two events will be sent: ABS_PRESSURE and EV_SYN.

The previous logic would discard in this case.  This is also
in preparation for simplier tools which do not report serial
numbers and in general will send small total events per
sync window.

Also, fixed comment since MSC_SERIAL events are not strictly
required.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Acked-by: Ping Cheng <pinglinux@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>