Rework rotation and scaling into tablet areas.
For each event, scale into the the range defined by topX, ..., then rotate
and invert the axes as required.

Don't mess with the topX, ... properties when rotating, the man page claims
they apply to the physical topX, ..., regardless of rotation.
Since we don't modify these anymore, rotateOneTool() is obsolete and
wcmRotateTablet() a lot simpler (to be removed out in the future).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Jan Steffens <jan.steffens@gmail.com>
Acked-by: Ping Cheng <pinglinux@gmail.com>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>