Partial support for Bamboo Touch MT-style interface.
Initial changes to support Bamboo Touch input device
reporting values using a touchpad-like event interface
and MT events for multitouch.  This change is backwards
compatible with original "serial #" style multiplexing of
touch data used by Tablet PC's and Bamboo Touch's.

This change allows a generic touchpad to function with
xf86-input-wacom; including basic 1 finger gestures such as
tap-to-click.

Tablet buttons do not work while finger is out-of-proximity.
They need to be rerouted to always-in-proximity PAD device (last
channel) to work with current framework.

MT events are not processed yet and its 2nd finger events need to be
rerouted to channel #2 if wacom gesture logic is to work.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>