Support tablet buttons without serial #'s.
The core of xf86-input-wacom strictly enforces buttons
on tools that are out-of-proximity must be cleared except
for the special case of the PAD device that is always
considered in proximity.

Generic tablets (non-wacom) and touchpads will send button
presses associated with tablet itself even if no tools
are reported as in proximity.

Work around this by forcing all non-styus button presses
to be routed to hard coded PAD channel and post multiple
channel events per sync window.  This is for generic
tablets only and wacom tablets continue to process events
the same way.

MT packets could be implemented using same concept and
routing to hard coded channel 0 or 1 based on finger #.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>