move storing of raw samples into wcmFilter filer
This helps abstract out filtering better and also allows
custom behavior on storing of samples as well; such as not
moving avg window if no X/Y values have changed in current
event cycle.

Intent of this change is refactor only (no behavior change).
Most off diff is moving logic over to wcmFilter.c.  There is minor
flow difference because RawFilter() and wcmCheckSuppress() needs
to be kept in wcmCommon.c, and so outside logic that stores samples,
but intent is same results.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Reviewed-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>