BTN_TOOL_FINGER no longer a valid type for pad tool
BTN_TOOL_FINGER is used by mouse driver for the first finger.
Before a new BTN_TOOL_ is added in input.h for pad tool, we
don't have a BTN_TOOL_. However, we still need to configure
those tablet buttons and strips/ring. Let's detect the actual
buttons we are going to receive to determine the existence
of a pad tool.

BTN_0 and BTN_FORWARD are the two unique BTN_s that we use for
tablet buttons in the kernel driver.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>