Fix release script for sourceforge shell service.
sf requires the creation of a shell before any operations, see
http://sourceforge.net/apps/trac/sourceforge/wiki/Shell%20service
The shell creation always requires a username, so take that from $USER.

Leave the shell open at the end of the script (i.e. don't forcibly shut it
down) in case the user needs to go in and do additional changes.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>