Set maximum tool button count to correct values
The code to set nbuttons for non-pad buttons was checking
for a lot of buttons that no current wacom tablets send
or that is understood by xf86-input-wacom.  Remove that
dead code.

The remaining logic seem to be trying to detect mouse
versus other tools (stylus mainly) and set button count
based on that.

Changed to look for BTN_TOOL_MOUSE to work with wider range
of tablets without special checks.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Acked-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>