Purge screen coordinate storage.
Input drivers monitoring screens doesn't work out well, all we need to do is
use xrandr to change the screens and our driver deals with wrong
information.

Note that this patch breaks KeepShape (if it worked before), shape
proportions should be exported as a property and be controlled by the
client instead.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>