Only delta the values for non-PAD devices.
We claim the pad is a relative device, but only because we had troubles with
the cursor landing at 0/0 in the past. Pretend we're still a relative
device, but post the values as absolute ones to the server.

Note that the device mode is to tell how the device _behaves_, it doesn't
specify how the device sends valuators. Thus, a relative device may only
post absolute coordinates and vice versa.

This applies to motion events only, button events for the scroll ring still
have the wrong valuator information.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>