Copy valuators passed into xf86Post*EventP.
X servers pre 1.9 used the data passed in directly and modified it in the
process. A call of e.g.
    int valuators[];
    xf86PostMotionEventP(..., valuators);
    xf86PostButtonEventP(..., valuators);

would thus post garbage data for the button event. Work around this by
copying the valuators before passing them in.
This is only needed for ABI < 11, server 1.9 duplicates the data passed in
before processing.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>