Fix segfault in case of unplug during PreInit.
If the device is removed between wcmHotplugOthers() and the end of PreInit()
of the first device, the close(fd) call at the end of PreInit("stylus") will
cause a SIGIO on the file descriptor.

The read call (wcmReadPacket(), usually on the last hotplugged device, e.g.
the pad) will then result in an error, causing this device to close all
associated devices through wcmDevProc. The parameter passed in is
essentially pInfo->dev. This leads to a NULL-pointer dereference, as
pInfo->dev is not yet set up for the first device since we're still in
PreInit. Call stack is something like this:

PreInit(stylus)
        wcmHotplugOthers()
              PreInit(eraser)
              PreInit(cursor)
              ...
        close(fd)
......... unplug device here.............
        wcmReadInput(pad)
              read error
              wcmDevProc(stylus->dev)
              NULL-pointer dereference
Server sets stylus->pInfo->fd

Segfault only happens when the device is unplugged when the pad is already
active but the stylus is still before the server assigns pInfo->dev.
Reproducible test-case: set a break point on the close() call in PreInit for
the stylus, then unplug the device before continuing.

This patch adds a check whether the device is already set up before calling
close. If not, skip it. The DEVICE_INIT following PreInit will see a ENODEV
on the fd and thus remove the device and its dependent devices anyway.

Reported-by: Camille Moncelier <moncelier@devlife.org>
Reported-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>