config: missing randr extension not detected
Problem introduced in commit 8f1dd625e00c182ec1b6a92b277214d1b8f88986.

  # Store the list of server defined optional extensions in REQUIRED_MODULES
  XORG_DRIVER_CHECK_EXT(RANDR, randrproto)

The role of the driver check macro is to supply the name of optional extension
to the PKG_CHECK_MODULES through the REQUIRED_MODULES variable.
By reversing the order of the statements the variable is empty.

The real fix would be to have 2 PKG_CHECK_MODULES statements, one to ensure
the server is present at the correct level and one to check the dependencies.

This won't be necessary as the RANDR extension is not optional,
it is always present. It can simply go next to the others.

All the xorg drivers have this flaw.

This undetected missing randr ext problem only happens with server 1.8 or prior,
as a new Requires.private statement has been added to server 1.9.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>