Don't close devices on read error.
If the devices were removed, the hotplugging code will take care of them and
remove them. If the devices had some other error happen to them, leave them
there and hope whatever it was comes back.

If a statically added device was unplugged, the user needs to trigger
DEVICE_OFF and DEVICE_ON manually to reopen (e.g. VT switch).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Ping Cheng <pinglinux@gmail.com>