Switch the pad to forced absolute mode.
The pad was in relative mode because we kept submitting x/y axes along with
the other data (and x/y was always 0/0). Hence the pointer always moved to
the screen origin when the pad was used.

With the recent changes, we only submit valuators 3, 4, 5 for the pad and
can thus switch the pad to absolute mode.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>