Add wraparound detection for the scroll ring.
The I4 scrollring appears as absolute axis (0 - 71) to us, but as
ring-shaped device to the user. Hence, a coordinate transition from 0..71 or
71..0 is common but must be interpreted as a negative or postive movement
direction, respectively.

Add code to detect this by simply assuming that if the direct delta of
old_x → x is larger than the indirect delta (across the axis range
boundary), a wraparound happened. In that casel, switch the delta used for
scroll wheel button emulation around.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>
Reviewed-by: Ping Cheng <pinglinux@gmail.com>